backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Phase 2 dependencies
tiktoken==0.5.2
networkx==3.2.1
javalang==0.13.0
tree-sitter==0.20.4
tree-sitter-javascript==0.21.4
tree-sitter-typescript==0.23.2

# Phase 3 dependencies
# Note: httpx already included for Groq client
# No additional dependencies needed to stay within constraints
